Fixing viewer from not refreshing properly when synching an unreal project with UnrealSync.
Fixing viewer from not refreshing properly when synching an unreal project with UnrealGameSync.

Synching with UGS and then building the unreal project causes a small compile, then the actual compilation starts, but the Dashboard was not updating the work status. This is a bandaid but it works for now. (Needs more cleanup, tried several things before landing on a fix)